fix(rocjitsu): make the fp_mode.h boundary the strict-rounding boundary
Scoping -frounding-math to rocjitsu_isa_<arch>_exec and tests/ left the header
boundary and the option boundary in different places. compute_unit.cpp reaches
fp_mode.h through alu_exceptions.h and is compiled without the option; today it
odr-uses only effective_omod, so no fenv-sensitive helper is emitted there, but
that is a property of the current call sites and nothing enforces it. A future
call to fma_f64 or finish_f64 from that file would emit an unflagged inline
definition into the same image, and since the linker keeps one of the copies it
could be the one every caller ends up in.

Split the helpers that never touch the host FP environment -- effective_omod,
effective_f16_omod, the finalize_omod_* family and the input modifiers -- into
fp_mode_no_fenv.h, and point alu_exceptions.h at that. What is left in fp_mode.h
is exactly the code whose correctness depends on the compiler not assuming a
rounding mode, so it can now demand the contract: RJ_STRICT_FP_ROUNDING_OPTIONS
carries -DROCJITSU_STRICT_FP_ROUNDING alongside the flag, and fp_mode.h #errors
without it. Including it from an unflagged target is a build failure naming the
header to use instead, rather than a silent miscompile.

fma_f16 stays on the strict side. It rounds to F16 in software, but its
intermediate is a double std::fma whose result is not always exact at that
width, so the ambient rounding mode can still reach it.

Fills in the two branches the variable was missing while here: /fp:strict for
MSVC, and a warning rather than silence for a compiler with no known spelling.

ctest 3689/3689.
